backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
pydantic==2.9.2
PyMongo==4.8.0

